Do not use `g_memdup2()` without GLib version check
It is only available since GLib 2.68, which is greater than 2.52
required in `configure.ac.in`, and greater than the version currently
used in the test container, so this prevents CI from running.

This change introduces a deprecation warning, but it is best to address
this in a future merge request, along with other build warnings.